* Ajustes de tipagens de repositório  - todos estavam com Id Integer na extensão do JPA, sendo que a classes está com Id Long
* Criação de findAll usuarios no controller

* Função atualizarUsuario mais robusta
 - Verificação de alteração a ser feita, evitando save desnecessário

 - Verificação de nome ou email já cadastrado

 * Mais testes unitários

+++ src/main/java/com/qqd/edicria/repositories/tabelasPrincipais/AutorReposritory.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AutorReposritory extends JpaRepository<Autor, Integer> {
+public interface AutorReposritory extends JpaRepository<Autor, Long> {
 }
